# Improve `JSON::stringify` performance