- Changed `stringify` to call static function `_stringify` directly, instead of creating JSON object
- Changed `colon` and `end_statement` from `String` to `const char *` to avoid extra allocations in each `_stringify` call
- Pass result `String` reference to each `_stringify` call to append to instead of allocating new `String` in each call

These changes make `JSON::stringify` around 2-3x faster in most cases

Compared with gdscript below:
```
func _ready() -> void:
	var d : Dictionary = {
		"int": 4,
		"float": 1.23,
		"string": "test",
		"array": [1,2,3,4,5],
		"bool": true,
		"null": null
	}
	var ia = [1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30]
	var sa = ["one", "two", "three", "four", "five", "six", "seven", "eight", "nine", "ten"]
	var s = "test"
	var f = 1.2345
	var i = 100
	var b = true
	var n = null
	
	run_test(d, "Dictionary (indent)", "\t")
	run_test(d, "Dictionary (no indent)")
	run_test(ia, "int Array (indent)", "\t")
	run_test(ia, "int Array (no indent)")
	run_test(sa, "String Array (indent)", "\t")
	run_test(sa, "String Array (no indent)")
	run_test(s, "String")
	run_test(f, "float")
	run_test(i, "int")
	run_test(b, "bool")
	run_test(n, "null")
	
func run_test(obj, obj_name : String, indent : String = ""):
	var start_ms = Time.get_ticks_msec()
	for i in 1000000:
		JSON.stringify(obj, indent)
	var end_ms = Time.get_ticks_msec()
	print("%s %dms" % [obj_name, end_ms - start_ms])
```

old
```
Dictionary (indent) 16457ms
Dictionary (no indent) 11121ms
int Array (indent) 17641ms
int Array (no indent) 7691ms
String Array (indent) 11741ms
String Array (no indent) 8103ms
String 1068ms
float 1000ms
int 575ms
bool 568ms
null 560ms
```

new
```
Dictionary (indent) 6185ms
Dictionary (no indent) 5786ms
int Array (indent) 4920ms
int Array (no indent) 4275ms
String Array (indent) 3760ms
String Array (no indent) 3544ms
String 508ms
float 643ms
int 233ms
bool 213ms
null 204ms
```

<!--
Please target the `master` branch in priority.

Relevant fixes are cherry-picked for stable branches as needed by maintainers.

To speed up the contribution process and avoid CI errors, please set up pre-commit hooks locally:
https://docs.godotengine.org/en/latest/contributing/development/code_style_guidelines.html
-->