@Mickeon 2024-06-30
That's insane. Does `stringify` have associated tests with it? If yes and they pass, this is a slam-dunk of a PR