@aaronp64 2024-06-30
There's some `JSONRPC` tests that use `JSON::stringify`, but I don't see any that test `stringify` directly.  I'll look into setting up some tests, looks like we already have some for `parse`