@aaronp64 2024-07-01
Added some tests and checked that they passed with both old and new code, but looks like there's some memory leaks in the test cases, looking into.  Anyone know how to run the memory leak checks locally on Windows, so I can check before pushing again?