@kus04e4ek 2024-07-01
> Anyone know how to run the memory leak checks locally on Windows, so I can check before pushing again?

Build with `use_asan=yes` and run the `.san` file